refactor(core): Removed unused code
