--- core/app/__tests__/main.test.ts
@@ -1,3 +1,0 @@
-import { test } from 'vitest';
-
-test("", () => { });
