--- core/modules/counter-plugin.js
@@ -1,1 +1,0 @@
-(()=>{var On=Object.create;var Dr=Object.defineProperty;var Un=Object.getOwnPropertyDescriptor;var Wn=Object.getOwnPropertyNames;var wn=Object.getPrototypeOf,bn=Object.prototype.hasOwnProperty;var Lr=(r,t)=>()=>(t||r((t={exports:{}}).exports,t),t.exports),Nn=(r,t)=>{for(var n in t)Dr(r,n,{get:t[n],enumerable:!0})},Sn=(r,t,n,e)=>{if(t&&typeof t=="object"||typeof t=="function")for(let o of Wn(t))!bn.call(r,o)&&o!==n&&Dr(r,o,{get:()=>t[o],enumerable:!(e=Un(t,o))||e.enumerable});return r};var L=(r,t,n)=>(n=r!=null?On(wn(r)):{},Sn(t||!r||!r.__esModule?Dr(n,"default",{value:r,enumerable:!0}):n,r));var j=Lr(l=>{"use strict";var Fn=l&&l.__spreadArray||function(r,t,n){if(n||arguments.length===2)for(var e=0,o=t.length,u;e<o;e++)(u||!(e in t))&&(u||(u=Array.prototype.slice.call(t,0,e)),u[e]=t[e]);return r.concat(u||Array.prototype.slice.call(t))};Object.defineProperty(l,"__esModule",{value:!0});l.dual=l.getEndomorphismMonoid=l.SK=l.hole=l.constVoid=l.constUndefined=l.constNull=l.constFalse=l.constTrue=l.unsafeCoerce=l.apply=l.getRing=l.getSemiring=l.getMonoid=l.getSemigroup=l.getBooleanAlgebra=void 0;l.identity=jr;l.constant=rr;l.flip=Hn;l.flow=dt;l.tuple=Bn;l.increment=kn;l.decrement=Kn;l.absurd=mt;l.tupled=zn;l.untupled=Jn;l.pipe=Gn;l.not=Zn;var Vn=function(r){return function(){return{meet:function(t,n){return function(e){return r.meet(t(e),n(e))}},join:function(t,n){return function(e){return r.join(t(e),n(e))}},zero:function(){return r.zero},one:function(){return r.one},implies:function(t,n){return function(e){return r.implies(t(e),n(e))}},not:function(t){return function(n){return r.not(t(n))}}}}};l.getBooleanAlgebra=Vn;var Cn=function(r){return function(){return{concat:function(t,n){return function(e){return r.concat(t(e),n(e))}}}}};l.getSemigroup=Cn;var Pn=function(r){var t=(0,l.getSemigroup)(r);return function(){return{concat:t().concat,empty:function(){return r.empty}}}};l.getMonoid=Pn;var Dn=function(r){return{add:function(t,n){return function(e){return r.add(t(e),n(e))}},zero:function(){return r.zero},mul:function(t,n){return function(e){return r.mul(t(e),n(e))}},one:function(){return r.one}}};l.getSemiring=Dn;var Ln=function(r){var t=(0,l.getSemiring)(r);return{add:t.add,mul:t.mul,one:t.one,zero:t.zero,sub:function(n,e){return function(o){return r.sub(n(o),e(o))}}}};l.getRing=Ln;var jn=function(r){return function(t){return t(r)}};l.apply=jn;function jr(r){return r}l.unsafeCoerce=jr;function rr(r){return function(){return r}}l.constTrue=rr(!0);l.constFalse=rr(!1);l.constNull=rr(null);l.constUndefined=rr(void 0);l.constVoid=l.constUndefined;function Hn(r){return function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return t.length>1?r(t[1],t[0]):function(e){return r(e)(t[0])}}}function dt(r,t,n,e,o,u,i,a,p){switch(arguments.length){case 1:return r;case 2:return function(){return t(r.apply(this,arguments))};case 3:return function(){return n(t(r.apply(this,arguments)))};case 4:return function(){return e(n(t(r.apply(this,arguments))))};case 5:return function(){return o(e(n(t(r.apply(this,arguments)))))};case 6:return function(){return u(o(e(n(t(r.apply(this,arguments))))))};case 7:return function(){return i(u(o(e(n(t(r.apply(this,arguments)))))))};case 8:return function(){return a(i(u(o(e(n(t(r.apply(this,arguments))))))))};case 9:return function(){return p(a(i(u(o(e(n(t(r.apply(this,arguments)))))))))}}}function Bn(){for(var r=[],t=0;t<arguments.length;t++)r[t]=arguments[t];return r}function kn(r){return r+1}function Kn(r){return r-1}function mt(r){throw new Error("Called `absurd` function which should be uncallable")}function zn(r){return function(t){return r.apply(void 0,t)}}function Jn(r){return function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return r(t)}}function Gn(r,t,n,e,o,u,i,a,p){switch(arguments.length){case 1:return r;case 2:return t(r);case 3:return n(t(r));case 4:return e(n(t(r)));case 5:return o(e(n(t(r))));case 6:return u(o(e(n(t(r)))));case 7:return i(u(o(e(n(t(r))))));case 8:return a(i(u(o(e(n(t(r)))))));case 9:return p(a(i(u(o(e(n(t(r))))))));default:{for(var f=arguments[0],x=1;x<arguments.length;x++)f=arguments[x](f);return f}}}l.hole=mt;var $n=function(r,t){return t};l.SK=$n;function Zn(r){return function(t){return!r(t)}}var Xn=function(){return{concat:function(r,t){return dt(r,t)},empty:jr}};l.getEndomorphismMonoid=Xn;var Qn=function(r,t){var n=typeof r=="number"?function(e){return e.length>=r}:r;return function(){var e=Array.from(arguments);return n(arguments)?t.apply(this,e):function(o){return t.apply(void 0,Fn([o],e,!1))}}};l.dual=Qn});var Er=Lr(v=>{"use strict";Object.defineProperty(v,"__esModule",{value:!0});v.eqDate=v.eqNumber=v.eqString=v.eqBoolean=v.eq=v.strictEqual=v.getStructEq=v.getTupleEq=v.Contravariant=v.getMonoid=v.getSemigroup=v.eqStrict=v.URI=v.contramap=v.tuple=v.struct=v.fromEquals=void 0;var Oe=j(),Ue=function(r){return{equals:function(t,n){return t===n||r(t,n)}}};v.fromEquals=Ue;var We=function(r){return(0,v.fromEquals)(function(t,n){for(var e in r)if(!r[e].equals(t[e],n[e]))return!1;return!0})};v.struct=We;var we=function(){for(var r=[],t=0;t<arguments.length;t++)r[t]=arguments[t];return(0,v.fromEquals)(function(n,e){return r.every(function(o,u){return o.equals(n[u],e[u])})})};v.tuple=we;var be=function(r,t){return(0,Oe.pipe)(r,(0,v.contramap)(t))},Ne=function(r){return function(t){return(0,v.fromEquals)(function(n,e){return t.equals(r(n),r(e))})}};v.contramap=Ne;v.URI="Eq";v.eqStrict={equals:function(r,t){return r===t}};var Se={equals:function(){return!0}},Fe=function(){return{concat:function(r,t){return(0,v.fromEquals)(function(n,e){return r.equals(n,e)&&t.equals(n,e)})}}};v.getSemigroup=Fe;var Ve=function(){return{concat:(0,v.getSemigroup)().concat,empty:Se}};v.getMonoid=Ve;v.Contravariant={URI:v.URI,contramap:be};v.getTupleEq=v.tuple;v.getStructEq=v.struct;v.strictEqual=v.eqStrict.equals;v.eq=v.Contravariant;v.eqBoolean=v.eqStrict;v.eqString=v.eqStrict;v.eqNumber=v.eqStrict;v.eqDate={equals:function(r,t){return r.valueOf()===t.valueOf()}}});var cn=Lr(c=>{"use strict";Object.defineProperty(c,"__esModule",{value:!0});c.ordDate=c.ordNumber=c.ordString=c.ordBoolean=c.ord=c.getDualOrd=c.getTupleOrd=c.between=c.clamp=c.max=c.min=c.geq=c.leq=c.gt=c.lt=c.equals=c.trivial=c.Contravariant=c.getMonoid=c.getSemigroup=c.URI=c.contramap=c.reverse=c.tuple=c.fromCompare=c.equalsDefault=void 0;var Xe=Er(),wr=j(),Qe=function(r){return function(t,n){return t===n||r(t,n)===0}};c.equalsDefault=Qe;var Ye=function(r){return{equals:(0,c.equalsDefault)(r),compare:function(t,n){return t===n?0:r(t,n)}}};c.fromCompare=Ye;var ro=function(){for(var r=[],t=0;t<arguments.length;t++)r[t]=arguments[t];return(0,c.fromCompare)(function(n,e){for(var o=0;o<r.length-1;o++){var u=r[o].compare(n[o],e[o]);if(u!==0)return u}return r[o].compare(n[o],e[o])})};c.tuple=ro;var to=function(r){return(0,c.fromCompare)(function(t,n){return r.compare(n,t)})};c.reverse=to;var no=function(r,t){return(0,wr.pipe)(r,(0,c.contramap)(t))},eo=function(r){return function(t){return(0,c.fromCompare)(function(n,e){return t.compare(r(n),r(e))})}};c.contramap=eo;c.URI="Ord";var oo=function(){return{concat:function(r,t){return(0,c.fromCompare)(function(n,e){var o=r.compare(n,e);return o!==0?o:t.compare(n,e)})}}};c.getSemigroup=oo;var uo=function(){return{concat:(0,c.getSemigroup)().concat,empty:(0,c.fromCompare)(function(){return 0})}};c.getMonoid=uo;c.Contravariant={URI:c.URI,contramap:no};c.trivial={equals:wr.constTrue,compare:(0,wr.constant)(0)};var ao=function(r){return function(t){return function(n){return n===t||r.compare(n,t)===0}}};c.equals=ao;var io=function(r){return function(t,n){return r.compare(t,n)===-1}};c.lt=io;var po=function(r){return function(t,n){return r.compare(t,n)===1}};c.gt=po;var co=function(r){return function(t,n){return r.compare(t,n)!==1}};c.leq=co;var fo=function(r){return function(t,n){return r.compare(t,n)!==-1}};c.geq=fo;var lo=function(r){return function(t,n){return t===n||r.compare(t,n)<1?t:n}};c.min=lo;var vo=function(r){return function(t,n){return t===n||r.compare(t,n)>-1?t:n}};c.max=vo;var so=function(r){var t=(0,c.min)(r),n=(0,c.max)(r);return function(e,o){return function(u){return n(t(u,o),e)}}};c.clamp=so;var xo=function(r){var t=(0,c.lt)(r),n=(0,c.gt)(r);return function(e,o){return function(u){return!(t(u,e)||n(u,o))}}};c.between=xo;c.getTupleOrd=c.tuple;c.getDualOrd=c.reverse;c.ord=c.Contravariant;function mo(r,t){return r<t?-1:r>t?1:0}var ft={equals:Xe.eqStrict.equals,compare:mo};c.ordBoolean=ft;c.ordString=ft;c.ordNumber=ft;c.ordDate=(0,wr.pipe)(c.ordNumber,(0,c.contramap)(function(r){return r.valueOf()}))});var br=L(j());var Yn=function(r,t,n){if(n||arguments.length===2)for(var e=0,o=t.length,u;e<o;e++)(u||!(e in t))&&(u||(u=Array.prototype.slice.call(t,0,e)),u[e]=t[e]);return r.concat(u||Array.prototype.slice.call(t))};function H(r){return r}var ht=H;function Tr(r,t,n,e,o,u,i,a,p){switch(arguments.length){case 1:return r;case 2:return function(){return t(r.apply(this,arguments))};case 3:return function(){return n(t(r.apply(this,arguments)))};case 4:return function(){return e(n(t(r.apply(this,arguments))))};case 5:return function(){return o(e(n(t(r.apply(this,arguments)))))};case 6:return function(){return u(o(e(n(t(r.apply(this,arguments))))))};case 7:return function(){return i(u(o(e(n(t(r.apply(this,arguments)))))))};case 8:return function(){return a(i(u(o(e(n(t(r.apply(this,arguments))))))))};case 9:return function(){return p(a(i(u(o(e(n(t(r.apply(this,arguments)))))))))}}}function W(r,t,n,e,o,u,i,a,p){switch(arguments.length){case 1:return r;case 2:return t(r);case 3:return n(t(r));case 4:return e(n(t(r)));case 5:return o(e(n(t(r))));case 6:return u(o(e(n(t(r)))));case 7:return i(u(o(e(n(t(r))))));case 8:return a(i(u(o(e(n(t(r)))))));case 9:return p(a(i(u(o(e(n(t(r))))))));default:{for(var f=arguments[0],x=1;x<arguments.length;x++)f=arguments[x](f);return f}}}var O=function(r,t){var n=typeof r=="number"?function(e){return e.length>=r}:r;return function(){var e=Array.from(arguments);return n(arguments)?t.apply(this,e):function(o){return t.apply(void 0,Yn([o],e,!1))}}};var re=function(r,t,n){if(n||arguments.length===2)for(var e=0,o=t.length,u;e<o;e++)(u||!(e in t))&&(u||(u=Array.prototype.slice.call(t,0,e)),u[e]=t[e]);return r.concat(u||Array.prototype.slice.call(t))};var B={_tag:"None"},k=function(r){return{_tag:"Some",value:r}},Hr=function(r){return r._tag==="Left"};var gt=function(r){return{_tag:"Left",left:r}},yt=function(r){return{_tag:"Right",right:r}};var Rr=function(r){return r[0]};var Br=function(r){return re([r[0]],r.slice(1),!0)};function K(r){return function(t,n){return r.chain(t,function(e){return r.map(n(e),function(){return e})})}}function Rt(r){return function(t){return Tr(t,r.fromEither)}}function _t(r,t){var n=Rt(r),e=K(t);return function(o,u){return e(o,n(u))}}function nr(r){return function(t,n){return r.map(t,function(){return n})}}function _r(r){var t=nr(r);return function(n){return t(n,void 0)}}var It=function(r){return{equals:function(t,n){return t===n||r(t,n)}}};var kr={equals:function(r,t){return r===t}};var Ru=kr.equals;var te=function(r){return function(t,n){return t===n||r(t,n)===0}},At=function(r){return{equals:te(r),compare:function(t,n){return t===n?0:r(t,n)}}};function ne(r,t){return r<t?-1:r>t?1:0}var Au={equals:kr.equals,compare:ne};var oe=Rr;var qt=oe;var Kr=function(r,t,n){if(n||arguments.length===2)for(var e=0,o=t.length,u;e<o;e++)(u||!(e in t))&&(u||(u=Array.prototype.slice.call(t,0,e)),u[e]=t[e]);return r.concat(u||Array.prototype.slice.call(t))},Ot=function(r){return r.length>0};var Ut=function(r){return function(t){return Kr([r],t,!0)}},zr=Ut,Wt=function(r){return function(t){return Kr(Kr([],t,!0),[r],!1)}},wt=Wt;var ae=Br;var ie=ae;var bt=function(r){return function(t){var n=Math.max(1,r);return n>=t.length?[ie(t),[]]:[W(t.slice(1,n),zr(Ar(t))),t.slice(n)]}};var Ar=qt,Jr=function(r){return r.slice(1)};var er={equals:function(r,t){return r===t}},Gr={equals:er.equals,compare:function(r,t){return r<t?-1:r>t?1:0}},Hu={equals:er.equals,compare:Gr.compare,top:1/0,bottom:-1/0};var pe={concat:function(r,t){return r-t}},St={concat:function(r,t){return r+t}},Ft={concat:function(r,t){return r*t}},Bu={concat:St.concat,empty:0},ku={concat:Ft.concat,empty:1},Ku={add:St.concat,zero:0,mul:Ft.concat,one:1,sub:pe.concat,degree:function(r){return 1},div:function(r,t){return r/t},mod:function(r,t){return r%t}};function Vt(r){return function(t){for(var n=0;n<t.length;n++)if(r(t[n]))return k(t[n]);return B}}var Ct=function(r){return function(t){return function(n){return n.reduce(function(e,o,u){return r.concat(e,t(u,o))},r.empty)}}};var Pt=function(r){var t=Ct(r);return function(n){return t(function(e,o){return n(o)})}};var Dt=function(r){return It(function(t,n){return t.length===n.length&&t.every(function(e,o){return r.equals(e,n[o])})})},Lt=function(r){return At(function(t,n){for(var e=t.length,o=n.length,u=Math.min(e,o),i=0;i<u;i++){var a=r.compare(t[i],n[i]);if(a!==0)return a}return Gr.compare(e,o)})};var $r=function(r){return r.length===0},Zr=Ot,Xr=zr;var Qr=wt;var jt=function(r){return r.length};var ve=function(r,t){for(var n=r.length,e=0;e<n&&t(r[e]);e++);return e};function Ht(r){return function(t){var n=xe(ve(t,r))(t),e=n[0],o=n[1];return{init:e,rest:o}}}function Yr(r){return Vt(r)}var or=function(r){return r.slice()};var rt=function(r){return function(t){return t.length<=1?or(t):t.slice().sort(r.compare)}},se=function(r,t,n){for(var e=[],o=Math.min(r.length,t.length),u=0;u<o;u++)e[u]=n(r[u],t[u]);return e};function ur(r,t){return t===void 0?function(n){return ur(n,r)}:se(r,t,function(n,e){return[n,e]})}var xe=function(r){return function(t){return r>=1&&Zr(t)?bt(r)(t):$r(t)?[or(t),[]]:[[],or(t)]}};var de=function(r){return function(t){return $r(t)?or(r):$r(r)?or(t):t.concat(r)}},tt=de;var ar=Pt;var nt=Dt,Bt=Lt;var w=B,S=k;var Te=function(r){return r._tag==="Left"?w:S(r.right)},kt=function(r,t){return W(r,Kt(t))},Re=function(r,t){return W(r,_e(t))};var ot="Option";var Kt=function(r){return function(t){return V(t)?w:S(r(t.value))}},zt={URI:ot,map:kt},Da=O(2,nr(zt)),La=_r(zt);var _e=function(r){return function(t){return V(t)||V(r)?w:S(t.value(r.value))}};var Ie=O(2,function(r,t){return V(r)?w:t(r.value)}),Jt={URI:ot,map:kt,ap:Re,chain:Ie};var ja=O(2,function(r,t){return V(r)?t():r});var Ae=Te,Me={URI:ot,fromEither:Ae};var V=function(r){return r._tag==="None"},Ee=function(r,t){return function(n){return V(n)?r():t(n.value)}};var Mr=Ee;var qe=function(r){return function(t){return V(t)?r():t.value}},ut=qe;var Ha=O(2,K(Jt)),Ba=O(2,_t(Me,Jt));var Gt=function(r){return r==null?w:S(r)};var Zt=ht;var vn=L(Er());var I={equals:function(r,t){return r===t}},Ce={concat:function(r,t){return r+t}},Pe="",Ja={concat:Ce.concat,empty:Pe},Xt={equals:I.equals,compare:function(r,t){return r<t?-1:r>t?1:0}};var Qt=gt,qr=yt,De=O(2,function(r,t){return M(r)?r:t(r.right)}),Yt=function(r,t){return W(r,je(t))},Le=function(r,t){return W(r,Be(t))};var at="Either";var je=function(r){return function(t){return M(t)?t:qr(r(t.right))}},rn={URI:at,map:Yt},ni=O(2,nr(rn)),ei=_r(rn);var He=function(r){return function(t){return M(t)?t:M(r)?r:qr(t.right(r.right))}},Be=He;var ke={URI:at,map:Yt,ap:Le,chain:De};var Ke={URI:at,fromEither:H};var M=Hr;var oi=O(2,K(ke));var ui={fromEither:Ke.fromEither};function tn(r){return r[0]}var z=tn;var lt=L(j());var ir=r=>"int"in r,pr=r=>"float"in r,cr=r=>"str"in r,fr=r=>"bool"in r,lr=r=>"list"in r,vr=r=>"obj"in r;var en=r=>({int:r});var U=L(Er());var sr=L(j());var Or={equals:function(r,t){return r===t}};var Je={concat:function(r,t){return r&&t}},Ge={concat:function(r,t){return r||t}},Ur={concat:Je.concat,empty:!0},on={concat:Ge.concat,empty:!1},si={equals:Or.equals,compare:function(r,t){return r<t?-1:r>t?1:0}};var an=(0,U.fromEquals)(([r,t],[n,e])=>I.equals(r,n)),it=(0,U.fromEquals)(([r,t],[n,e])=>I.equals(r,n)?pt.equals(t,e):!1),_i=(0,U.fromEquals)((r,t)=>(0,sr.pipe)(ur(r)(t),ar(Ur)(([n,e])=>it.equals(n,e)))),pn=(0,U.fromEquals)((r,t)=>r.length===0||t.length===0?!1:(0,sr.pipe)(tt(r)(t),n=>n,rt(ct),Wr(an),ar(on)(n=>jt(n)!=1))),pt=(0,U.fromEquals)((r,t)=>ir(r)&&ir(t)?er.equals(r.int,t.int):pr(r)&&pr(t)?er.equals(r.float,t.float):cr(r)&&cr(t)?I.equals(r.str,t.str):fr(r)&&fr(t)?Or.equals(r.bool,t.bool):lr(r)&&lr(t)?nt(pt).equals(r.list,t.list):vr(r)&&vr(t)?nt(it).equals(Array.from(r.obj.values()),Array.from(t.obj.values())):!1),Ii=(0,U.fromEquals)((r,t)=>I.equals(z(r),z(t))),$e=(0,U.fromEquals)(({kind:r,kids:t,text:n,attrs:e},{kind:o,kids:u,text:i,attrs:a})=>I.equals(r,o)&&(0,sr.pipe)(t,ur(u),ar(Ur)(([p,f])=>$e.equals(p,f)))&&(n===null&&i===null||n!==null&&i!==null&&I.equals(n,i))&&(0,sr.pipe)(e,ur(a),ar(Ur)(([p,f])=>Ze.equals(p,f)))),Ze=(0,U.fromEquals)((r,t)=>"id"in r&&"id"in t&&I.equals(r.id,t.id)||"class"in r&&"class"in t&&I.equals(r.class,t.class)||"style"in r&&"style"in t&&I.equals(r.style,t.style)||"type"in r&&"type"in t&&I.equals(r.type,t.type)||"checked"in r&&"checked"in t&&Or.equals(r.checked,t.checked)||"onClick"in r&&"onClick"in t&&un.equals(r.onClick,t.onClick)||"onInput"in r&&"onInput"in t&&un.equals(r.onInput,t.onInput)||"emitInput"in r&&"emitInput"in t&&I.equals(r.emitInput,t.emitInput)||"src"in r&&"src"in t&&I.equals(r.src,t.src)),un=(0,U.fromEquals)(({msg:[r,t]},{msg:[n,e]})=>I.equals(r,n)&&pt.equals(t,e));var fn=L(cn());var F=class r{_kind="frag";_kids=[];_attrs=[];_text=null;kind(t){return this._kind=t,this}kids(...t){return this._kids=t,this}attrs(...t){return this._attrs=t,this}text(t){return this._text=t,this}build(){return{kind:this._kind,kids:this._kids.map(t=>t instanceof r?t.build():t),attrs:this._attrs,text:this._text}}};var ln=r=>t=>(0,lt.pipe)(t[r],Gt,Mr(()=>w,n=>ho(n)?S(n):w));var ho=r=>ir(r)||pr(r)||fr(r)||cr(r)||lr(r)?!0:vr(r);var Wr=r=>t=>Zr(t)?(0,lt.pipe)(Ar(t),n=>{let e=Ht(a=>r.equals(n,a))(Jr(t)),o=e.init,u=e.rest;return Xr(Xr(n)(o))(Wr(r)(u))}):[];var ct=(0,fn.fromCompare)((r,t)=>Xt.compare(z(r),z(t))),ji=Bt(ct);var Ji=(0,vn.fromEquals)(({event:r,module:t},{event:n,module:e})=>I.equals(r,n)&&I.equals(t,e)),Nr=class{ui;uiModifications=[];state;stateModifications=[];events;eventModifications=[];newEventHandlers=[];constructor(t){this.ui=t.ui,this.state=t.state,this.events=t.events}traverseUI(t,n){return n(t)?S(t):(0,br.pipe)(t.kids,Yr(n))}findUI(t){return(0,br.pipe)(this.traverseUI(this.ui,t),ut(()=>{}))}addUI(t,n){let e={f:n,field:"kids",g:o=>Qr(t)(o)};return this.uiModifications.push(e),this}modifyUI(t,n,e){let o={f:t,field:n,g:e};return this.uiModifications.push(o),this}removeUI(t){let n={f:t};return this.uiModifications.push(n),this}findField(t){return(0,br.pipe)(this.state,ln(t),Mr(()=>{},Zt))}build(){return{newEventHandlers:this.newEventHandlers,uiModifications:this.uiModifications,stateModifications:this.stateModifications,eventModifications:this.eventModifications}}};var Pr={};Nn(Pr,{DAttrs:()=>_n,DCoreModification:()=>cu,DEvent:()=>Mn,DEventHandler:()=>qn,DHtml:()=>Q,DHtmlKind:()=>In,DInitDecoder:()=>uu,DIns:()=>gr,DInsArr:()=>pu,DMap:()=>D,DMapArr:()=>ou,DMsg:()=>xt,DNode:()=>iu,DUpdateDecoder:()=>An,DValue:()=>X,DValueArr:()=>eu,DViewDecoder:()=>au});var d=function(){var r=function(t,n){return r=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(e,o){e.__proto__=o}||function(e,o){for(var u in o)Object.prototype.hasOwnProperty.call(o,u)&&(e[u]=o[u])},r(t,n)};return function(t,n){if(typeof n!="function"&&n!==null)throw new TypeError("Class extends value "+String(n)+" is not a constructor or null");r(t,n);function e(){this.constructor=t}t.prototype=n===null?Object.create(n):(e.prototype=n.prototype,new e)}}(),Sr=function(){return Sr=Object.assign||function(r){for(var t,n=1,e=arguments.length;n<e;n++){t=arguments[n];for(var o in t)Object.prototype.hasOwnProperty.call(t,o)&&(r[o]=t[o])}return r},Sr.apply(this,arguments)};var $=Qt,E=function(r,t,n){return $([{value:r,context:t,message:n}])},_=qr,m=function(){function r(t,n,e,o){this.name=t,this.is=n,this.validate=e,this.encode=o,this.decode=this.decode.bind(this)}return r.prototype.pipe=function(t,n){var e=this;return n===void 0&&(n="pipe(".concat(this.name,", ").concat(t.name,")")),new r(n,t.is,function(o,u){var i=e.validate(o,u);return M(i)?i:t.validate(i.right,u)},this.encode===g&&t.encode===g?g:function(o){return e.encode(t.encode(o))})},r.prototype.asDecoder=function(){return this},r.prototype.asEncoder=function(){return this},r.prototype.decode=function(t){return this.validate(t,[{key:"",type:this,actual:t}])},r}();var g=function(r){return r};function yo(r){return r.displayName||r.name||"<function".concat(r.length,">")}function J(r,t,n,e){for(var o=r.length,u=Array(o+1),i=0;i<o;i++)u[i]=r[i];return u[o]={key:t,type:n,actual:e},u}function dr(r,t){for(var n=t.length,e=0;e<n;e++)r.push(t[e])}var xr=Object.prototype.hasOwnProperty;function To(r){return Object.keys(r).map(function(t){return"".concat(t,": ").concat(r[t].name)}).join(", ")}function Fr(r){for(var t=0;t<r.length;t++)if(r[t].encode!==g)return!1;return!0}function Ro(r){return"{ ".concat(To(r)," }")}function _o(r){return"("+r.map(function(t){return t.name}).join(" | ")+")"}function sn(r,t){for(var n=!0,e=!0,o=!G.is(r),u=0,i=t;u<i.length;u++){var a=i[u];a!==r&&(n=!1),G.is(a)&&(e=!1)}if(n)return r;if(e)return t[t.length-1];for(var p={},f=0,x=t;f<x.length;f++){var a=x[f];for(var h in a)(!xr.call(p,h)||o||a[h]!==r[h])&&(p[h]=a[h])}return p}function Io(r){return r.length>0}var q={};function dn(r,t){for(var n=[],e=0,o=r;e<o.length;e++){var u=o[e];t.indexOf(u)!==-1&&n.push(u)}return n}function Ao(r,t){if(r===q)return t;if(t===q)return r;var n=Object.assign({},r);for(var e in t)if(xr.call(r,e)){var o=dn(r[e],t[e]);if(Io(o))n[e]=o;else{n=q;break}}else n[e]=t[e];return n}function Mo(r,t){if(r===q||t===q)return q;var n=q;for(var e in r)if(xr.call(t,e)){var o=dn(r[e],t[e]);o.length===0&&(n===q&&(n={}),n[e]=r[e].concat(t[e]))}return n}function Eo(r){return r._tag==="LiteralType"}function qo(r){return r._tag==="InterfaceType"}function Oo(r){return r._tag==="StrictType"}function Uo(r){return r._tag==="ExactType"}function Wo(r){return r._tag==="RefinementType"}function wo(r){return r._tag==="IntersectionType"}function bo(r){return r._tag==="UnionType"}function No(r){return r._tag==="RecursiveType"}var vt=[];function C(r){if(vt.indexOf(r)!==-1)return q;if(qo(r)||Oo(r)){var t=q;for(var n in r.props){var e=r.props[n];Eo(e)&&(t===q&&(t={}),t[n]=[e.value])}return t}else{if(Uo(r)||Wo(r))return C(r.type);if(wo(r))return r.types.reduce(function(u,i){return Ao(u,C(i))},q);if(bo(r))return r.types.slice(1).reduce(function(u,i){return Mo(u,C(i))},C(r.types[0]));if(No(r)){vt.push(r);var o=C(r.type);return vt.pop(),o}}return q}function So(r){var t=C(r[0]),n=Object.keys(t),e=r.length,o=function(f){for(var x=t[f].slice(),h=[t[f]],T=1;T<e;T++){var b=r[T],yr=C(b),N=yr[f];if(N===void 0)return"continue-keys";if(N.some(function(Y){return x.indexOf(Y)!==-1}))return"continue-keys";x.push.apply(x,N),h.push(N)}return{value:[f,h]}};r:for(var u=0,i=n;u<i.length;u++){var a=i[u],p=o(a);if(typeof p=="object")return p.value;switch(p){case"continue-keys":continue r}}}var Fo=function(r){d(t,r);function t(){var n=r.call(this,"null",function(e){return e===null},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="NullType",n}return t}(m);var mn=new Fo,Vo=function(r){d(t,r);function t(){var n=r.call(this,"undefined",function(e){return e===void 0},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="UndefinedType",n}return t}(m);var xn=new Vo,Co=function(r){d(t,r);function t(){var n=r.call(this,"void",xn.is,xn.validate,g)||this;return n._tag="VoidType",n}return t}(m);var Zi=new Co,Po=function(r){d(t,r);function t(){var n=r.call(this,"unknown",function(e){return!0},_,g)||this;return n._tag="UnknownType",n}return t}(m);var Xi=new Po,Do=function(r){d(t,r);function t(){var n=r.call(this,"string",function(e){return typeof e=="string"},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="StringType",n}return t}(m);var R=new Do,Lo=function(r){d(t,r);function t(){var n=r.call(this,"number",function(e){return typeof e=="number"},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="NumberType",n}return t}(m);var mr=new Lo,jo=function(r){d(t,r);function t(){var n=r.call(this,"bigint",function(e){return typeof e=="bigint"},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="BigIntType",n}return t}(m);var Qi=new jo,Ho=function(r){d(t,r);function t(){var n=r.call(this,"boolean",function(e){return typeof e=="boolean"},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="BooleanType",n}return t}(m);var st=new Ho,Bo=function(r){d(t,r);function t(){var n=r.call(this,"UnknownArray",Array.isArray,function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="AnyArrayType",n}return t}(m);var Vr=new Bo,ko=function(r){d(t,r);function t(){var n=r.call(this,"UnknownRecord",function(e){return e!==null&&typeof e=="object"&&!Array.isArray(e)},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="AnyDictionaryType",n}return t}(m);var G=new ko;var Ko=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.value=i,a._tag="LiteralType",a}return t}(m);function s(r,t){t===void 0&&(t=JSON.stringify(r));var n=function(e){return e===r};return new Ko(t,n,function(e,o){return n(e)?_(r):E(e,o)},g,r)}var Yi=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.keys=i,a._tag="KeyofType",a}return t}(m);var zo=function(r){d(t,r);function t(n,e,o,u,i,a){var p=r.call(this,n,e,o,u)||this;return p.type=i,p.predicate=a,p._tag="RefinementType",p}return t}(m);function Jo(r,t,n){return Tn(r,t,n)}var rp=Jo(mr,function(r){return Number.isInteger(r)},"Int"),hn=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.runDefinition=i,a._tag="RecursiveType",a}return t}(m);Object.defineProperty(hn.prototype,"type",{get:function(){return this.runDefinition()},enumerable:!0,configurable:!0});function Cr(r,t){var n,e=function(){return n||(n=t(o),n.name=r),n},o=new hn(r,function(u){return e().is(u)},function(u,i){return e().validate(u,i)},function(u){return e().encode(u)},e);return o}var Go=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.type=i,a._tag="ArrayType",a}return t}(m);function A(r,t){return t===void 0&&(t="Array<".concat(r.name,">")),new Go(t,function(n){return Vr.is(n)&&n.every(r.is)},function(n,e){var o=Vr.validate(n,e);if(M(o))return o;for(var u=o.right,i=u.length,a=u,p=[],f=0;f<i;f++){var x=u[f],h=r.validate(x,J(e,String(f),r,x));if(M(h))dr(p,h.left);else{var T=h.right;T!==x&&(a===u&&(a=u.slice()),a[f]=T)}}return p.length>0?$(p):_(a)},r.encode===g?g:function(n){return n.map(r.encode)},r)}var $o=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.props=i,a._tag="InterfaceType",a}return t}(m);function y(r,t){t===void 0&&(t=Ro(r));var n=Object.keys(r),e=n.map(function(u){return r[u]}),o=n.length;return new $o(t,function(u){if(G.is(u)){for(var i=0;i<o;i++){var a=n[i],p=u[a];if(p===void 0&&!xr.call(u,a)||!e[i].is(p))return!1}return!0}return!1},function(u,i){var a=G.validate(u,i);if(M(a))return a;for(var p=a.right,f=p,x=[],h=0;h<o;h++){var T=n[h],b=f[T],yr=e[h],N=yr.validate(b,J(i,T,yr,b));if(M(N))dr(x,N.left);else{var Y=N.right;(Y!==b||Y===void 0&&!xr.call(f,T))&&(f===p&&(f=Sr({},p)),f[T]=Y)}}return x.length>0?$(x):_(f)},Fr(e)?g:function(u){for(var i=Sr({},u),a=0;a<o;a++){var p=n[a],f=e[a].encode;f!==g&&(i[p]=f(u[p]))}return i},r)}var tp=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.props=i,a._tag="PartialType",a}return t}(m);var np=function(r){d(t,r);function t(n,e,o,u,i,a){var p=r.call(this,n,e,o,u)||this;return p.domain=i,p.codomain=a,p._tag="DictionaryType",p}return t}(m);var gn=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.types=i,a._tag="UnionType",a}return t}(m);function Z(r,t){t===void 0&&(t=_o(r));var n=So(r);if(n!==void 0&&r.length>0){var e=n[0],o=n[1],u=o.length,i=function(a){for(var p=0;p<u;p++)if(o[p].indexOf(a)!==-1)return p};return new tu(t,function(a){if(G.is(a)){var p=i(a[e]);return p!==void 0?r[p].is(a):!1}return!1},function(a,p){var f=G.validate(a,p);if(M(f))return f;var x=f.right,h=i(x[e]);if(h===void 0)return E(a,p);var T=r[h];return T.validate(x,J(p,String(h),T,x))},Fr(r)?g:function(a){var p=i(a[e]);if(p===void 0)throw new Error("no codec found to encode value in union codec ".concat(t));return r[p].encode(a)},r,e)}else return new gn(t,function(a){return r.some(function(p){return p.is(a)})},function(a,p){for(var f=[],x=0;x<r.length;x++){var h=r[x],T=h.validate(a,J(p,String(x),h,a));if(M(T))dr(f,T.left);else return _(T.right)}return $(f)},Fr(r)?g:function(a){for(var p=0,f=r;p<f.length;p++){var x=f[p];if(x.is(a))return x.encode(a)}throw new Error("no codec found to encode value in union type ".concat(t))},r)}var Zo=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.types=i,a._tag="IntersectionType",a}return t}(m);function yn(r,t){t===void 0&&(t="(".concat(r.map(function(e){return e.name}).join(" & "),")"));var n=r.length;return new Zo(t,function(e){return r.every(function(o){return o.is(e)})},r.length===0?_:function(e,o){for(var u=[],i=[],a=0;a<n;a++){var p=r[a],f=p.validate(e,J(o,String(a),p,e));M(f)?dr(i,f.left):u.push(f.right)}return i.length>0?$(i):_(sn(e,u))},r.length===0?g:function(e){return sn(e,r.map(function(o){return o.encode(e)}))},r)}var Xo=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.types=i,a._tag="TupleType",a}return t}(m);function P(r,t){t===void 0&&(t="[".concat(r.map(function(e){return e.name}).join(", "),"]"));var n=r.length;return new Xo(t,function(e){return Vr.is(e)&&e.length===n&&r.every(function(o,u){return o.is(e[u])})},function(e,o){var u=Vr.validate(e,o);if(M(u))return u;for(var i=u.right,a=i.length>n?i.slice(0,n):i,p=[],f=0;f<n;f++){var x=i[f],h=r[f],T=h.validate(x,J(o,String(f),h,x));if(M(T))dr(p,T.left);else{var b=T.right;b!==x&&(a===i&&(a=i.slice()),a[f]=b)}}return p.length>0?$(p):_(a)},Fr(r)?g:function(e){return r.map(function(o,u){return o.encode(e[u])})},r)}var ep=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.type=i,a._tag="ReadonlyType",a}return t}(m);var op=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.type=i,a._tag="ReadonlyArrayType",a}return t}(m);var up=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.type=i,a._tag="ExactType",a}return t}(m);var Qo=function(r){d(t,r);function t(){var n=r.call(this,"Function",function(e){return typeof e=="function"},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="FunctionType",n}return t}(m);var hr=new Qo,Yo=function(r){d(t,r);function t(){var n=r.call(this,"never",function(e){return!1},function(e,o){return E(e,o)},function(){throw new Error("cannot encode never")})||this;return n._tag="NeverType",n}return t}(m);var ap=new Yo,ru=function(r){d(t,r);function t(){var n=r.call(this,"any",function(e){return!0},_,g)||this;return n._tag="AnyType",n}return t}(m);var ip=new ru;function Tn(r,t,n){return n===void 0&&(n="(".concat(r.name," | ").concat(yo(t),")")),new zo(n,function(e){return r.is(e)&&t(e)},function(e,o){var u=r.validate(e,o);if(M(u))return u;var i=u.right;return t(i)?_(i):E(i,o)},r.encode,r,t)}var pp=Tn(mr,Number.isInteger,"Integer"),tu=function(r){d(t,r);function t(n,e,o,u,i,a){var p=r.call(this,n,e,o,u,i)||this;return p.tag=a,p}return t}(gn);var nu=function(r){d(t,r);function t(){var n=r.call(this,"object",function(e){return e!==null&&typeof e=="object"},function(e,o){return n.is(e)?_(e):E(e,o)},g)||this;return n._tag="ObjectType",n}return t}(m);var cp=new nu;var fp=function(r){d(t,r);function t(n,e,o,u,i){var a=r.call(this,n,e,o,u)||this;return a.props=i,a._tag="StrictType",a}return t}(m);var X=Cr("DValue",()=>Z([y({int:mr}),y({float:mr}),y({bool:st}),y({str:R}),y({list:A(X)}),y({obj:A(P([R,X]))})])),eu=A(X),xt=y({msg:P([R,X])}),_n=Z([y({id:R}),y({class:R}),y({style:R}),y({onClick:xt}),y({onInput:xt}),y({emitInput:R}),y({src:R}),y({type:R}),y({checked:st})]),In=Z([s("div"),s("p"),s("h1"),s("h2"),s("h3"),s("h4"),s("h5"),s("h6"),s("span"),s("section"),s("article"),s("aside"),s("audio"),s("b"),s("br"),s("button"),s("code"),s("em"),s("fieldset"),s("form"),s("img"),s("input"),s("label"),s("link"),s("li"),s("menu"),s("nav"),s("ol"),s("option"),s("select"),s("style"),s("svg"),s("table"),s("td"),s("th"),s("ul"),s("video"),s("frag"),s("text"),s("script"),s("tr"),s("tbody"),s("main")]),Q=Cr("DHtml",()=>y({kind:In,kids:A(Q),text:Z([R,mn]),attrs:A(_n)})),D=A(P([R,X])),ou=A(D),An=A(P([R,D])),uu=An,au=A(P([R,Q])),Mn=y({event:R,module:R}),En=r=>Cr("DNodeImpl",()=>y({id:R,kids:A(En(r))})),iu=r=>yn([En(r),r]),gr=r=>Z([y({f:hr},"DRemIns"),y({f:hr,field:R,g:hr},"RModIns")]),qn=y({handler:hr,module:R}),pu=r=>A(gr(r)),cu=y({uiModifications:A(gr(Q)),stateModifications:A(gr(D)),eventModifications:A(gr(Mn)),newEventHandlers:A(P([R,qn]))});var{DViewDecoder:fu,DInitDecoder:lu,DUpdateDecoder:vu}=Pr,yp={init:lu,view:fu,update:vu,plugin_init:D,plugin_view:Q,plugin_update:D,get_plugins:A(R)},Tp={pluginAssets:[],root:document.body,client:(r,t)=>new Promise(()=>{}),log:{info:console.log,error:console.error},listen:(r,t)=>new Promise(n=>n(void 0)),emit:(r,t)=>new Promise(n=>n()),getPluginPaths:new Promise(r=>r([])),pluginInstallers:[r=>new Promise(t=>t(void 0))]};var su="CounterPlugin";window.plugins.set(su,{init:async r=>{let t=new Nr(r);return t.addUI(new F().kind("div").kids(new F().kind("button").attrs({onClick:{msg:["counter-event",en(1)]}}).text(`Count: ${t.findField("count")}`)).build(),n=>!0).build()}});})();
