--- core/modules/trivial.js
@@ -1,10 +1,0 @@
-window.plugins.set("trivial", {
-  init: async (_) => {
-    return {
-      uiModifications: [],
-      stateModifications: [],
-      eventModifications: [],
-      newEventHandlers: [],
-    };
-  },
-});
