+++ libs/js-utils/lib/Html.ts
@@ -2,6 +2,6 @@
 import type { TAttr } from "./TAttr";
 import type { UIInstruction } from "./UIInstruction";
 
-export type Html = { "text": string } | { "div": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "p": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "h1": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "h2": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "h3": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "h4": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "h5": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "h6": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "span": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "section": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "article": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "aside": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "audio": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "b": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "br": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "button": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "code": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "em": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "fieldset": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "form": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "img": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "input": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "label": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "link": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "li": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "menu": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "nav": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "ol": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "option": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "select": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "style": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "svg": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "table": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "td": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "th": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "tr": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "ul": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "video": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "frag": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "script": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "tbody": { kids: Array<Html>, attrs: Array<TAttr>, } } | { "main": { kids: Array<Html>, attrs: Array<TAttr>, } };
+export type Html = { "div": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "p": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "h1": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "h2": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "h3": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "h4": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "h5": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "h6": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "span": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "section": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "article": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "aside": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "audio": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "b": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "br": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "button": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "code": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "em": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "fieldset": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "form": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "img": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "input": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "label": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "link": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "li": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "menu": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "nav": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "ol": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "option": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "select": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "style": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "svg": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "table": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "td": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "th": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "tr": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "ul": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "video": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "frag": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "script": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "tbody": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } } | { "main": { kids: Array<Html>, attrs: Array<TAttr>, text: string | null, } };
 
 export type TUIInstruction = { op: "noOp" | { "addAttr": { field: string, attr: TAttr, } } | { "setAttr": { field: string, attr: TAttr, } } | { "remAttr": { field: string, } } | { "add": { ui: Html, id: string | null, class: string | null, } } | { "rem": { id: string | null, class: string | null, } } | { "addAttrPred": { field: string, id: string | null, class: string | null, attr: TAttr, } } | { "setAttrPred": { field: string, id: string | null, class: string | null, attr: TAttr, } } | { "remAttrPred": { id: string | null, class: string | null, field: string, } } | { "textAttrPred": { id: string | null, class: string | null, text: string, } } | { "then": { fst: UIInstruction, snd: UIInstruction, } }, };
